requirements.txt
streamlit
opencv-python-headless
pytesseract
googletrans==4.0.0-rc1
gtts==2.2.3
Pillow==8.4.0
